% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit theorem for a Poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of trials}

\item{iter}{number of iterations}

\item{lambda}{mean with time}

\item{...}{...}
}
\value{
graph of the Poisson distribution
}
\description{
Central Limit theorem for a Poisson distribution
}
\examples{
mycltp(n=10,iter=10000, lambda = 4)
}
