% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{area from to (-inf, a)}
}
\value{
graph and area
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
