% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats}

\item{gamma}{Probability threshold between 0 and 1}

\item{p}{probability of passengers showing up for flight}
}
\value{
printed list and 2 plots
}
\description{
ntickets
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
